tests/test_nas.py
import unittest
from unittest.mock import MagicMock

# Assuming the existence of an AbstractKVStore class and PathManagementMixin as designed.
from plugins.nas import PathManagementMixin
from store import AbstractKVStore

class EnhancedKVStore(AbstractKVStore, PathManagementMixin):
    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        PathManagementMixin.__init__(self, *args, **kwargs)

class TestPathManagementMixin(unittest.TestCase):
    def setUp(self):
        # Mocking the AbstractKVStore to focus on PathManagementMixin functionality.
        self.kv_store = EnhancedKVStore(backup_dir="test_backups", use_backup=False)
        self.kv_store._edit_key = MagicMock()
        self.kv_store._get_object = MagicMock()
        self.kv_store.create_store = MagicMock()

    def test_add_or_update_path(self):
        self.kv_store.add_or_update_path("myApp", "prod", "posix", "/var/app/prod")
        self.kv_store._edit_key.assert_called()

    def test_get_path(self):
        self.kv_store._get_object.return_value = {"prod": {"posix": "/var/app/prod"}}
        path = self.kv_store.get_path("myApp", "prod", "posix")
        self.assertEqual(path, "/var/app/prod")

    def test_update_paths_object(self):
        new_paths = {"dev": {"posix": "/var/app/dev"}}
        self.kv_store.update_paths_object("myApp", new_paths)
        self.kv_store._edit_key.assert_called_with("paths", "myApp", paths=new_paths)

    def test_edit_specific_path(self):
        self.kv_store.edit_specific_path("myApp", "prod", "posix", "/new/path")
        self.kv_store._edit_key.assert_called()

    def test_get_all_paths(self):
        self.kv_store._get_object.return_value = {"myApp": {"prod": {"posix": "/var/app/prod"}}}
        all_paths = self.kv_store.get_all_paths()
        self.assertIn("myApp", all_paths)

if __name__ == "__main__":
    unittest.main()
